Use drive.file scope for authentication.
Once we get a token, we hold it in a temporary session variable and
direct the user to a file picker with the team folder preselected as the
parent. The user can select the current year's puzzle folder to grant
the Big Board access to that folder, allowing us to do our usual check
that the folder is in fact readable by the user before permitting them
to enter.

This breaks puzzle creation, which will be handled separately by using a
service account to read the template and create new puzzle documents
once the user has authenticated.