change lowpriority icon
The low priority icon was set as meh-o, which makes sense because meh is low priority, but it was very ugly.  Also I couldn't find "meh-oh" anywhere, so I don't know where it was pulling that from.  This seems to be where all the other icons are defined so I thought maybe I could define this one here too, but honestly I put this at like a 15% chance of success